examples: refactor the tensorflow-backdoor-poisoning example
Refactors the tensorflow-backdoor-poisoning demo to use the new, common approach designed for the
examples folder. This includes the following changes:

- Moved any new custom task plugins from tensorflow-backdoor-poisoning folder to the root of the
  examples folder.
- Moved the MLproject file into the src folder.
- Refactored the task plugins and entrypoint scripts to remove the need for if/else logic at the
  top of the scripts.
- Pruned ununsed task plugins and entrypoint scripts from the example.
- Updated the documentation in examples/tensorflow-backdoor-poisoning/README.md with instructions
  on how the user should prepare their environment for running the examples.
- Modified the demo Jupyter notebooks for tensorflow-backdoor-poisoning to use the new scripts.
- Removed any files from the tensorflow-backdoor-poisoning folder that were no longer being used
  after implementing the above changes.
