+++ examples/task-plugins/dioptra_custom/custom_poisoning_plugins/data_tensorflow.py
