+++ examples/task-plugins/dioptra_custom/custom_poisoning_plugins/defenses_image_preprocessing.py
