+++ examples/tensorflow-backdoor-poisoning/src/gaussian_augmentation.py
@@ -74,13 +74,6 @@ def _coerce_int_to_bool(ctx, param, value):
 
 @click.command()
 @click.option(
-    "--data-dir",
-    type=click.Path(
-        exists=True, file_okay=False, dir_okay=True, resolve_path=True, readable=True
-    ),
-    help="Root directory for NFS mounted datasets (in container)",
-)
-@click.option(
     "--image-size",
     type=click.STRING,
     callback=_coerce_comma_separated_ints,
@@ -165,7 +158,6 @@ def _coerce_int_to_bool(ctx, param, value):
     default=-1,
 )
 def guassian_augmentation(
-    data_dir,
     image_size,
     def_tar_name,
     def_data_dir,
@@ -181,11 +173,9 @@ def guassian_augmentation(
     dataset_name,
     seed,
 ):
-
     LOGGER.info(
         "Execute MLFlow entry point",
         entry_point="guassian_augmentation",
-        data_dir=data_dir,
         image_size=image_size,
         def_tar_name=def_tar_name,
         def_data_dir=def_data_dir,
@@ -201,20 +191,11 @@ def guassian_augmentation(
         seed=seed,
     )
 
-    if load_dataset_from_mlruns:
-        data_dir = Path.cwd() / "dataset" / dataset_name
-        data_tar_name = dataset_tar_name
-        data_tar_path = download_image_archive(
-            run_id=dataset_run_id, archive_path=data_tar_name
-        )
-        with tarfile.open(data_tar_path, "r:gz") as f:
-            f.extractall(path=(Path.cwd() / "dataset"))
-
     with mlflow.start_run() as active_run:  # noqa: F841
         flow: Flow = init_guassian_augmentation_flow()
         state = flow.run(
             parameters=dict(
-                testing_dir=Path(data_dir),
+                testing_dir=(Path.cwd() / dataset_name).resolve(),
                 image_size=image_size,
                 def_tar_name=def_tar_name,
                 def_data_dir=(Path.cwd() / def_data_dir).resolve(),
@@ -226,27 +207,12 @@ def guassian_augmentation(
                 gaussian_augmentation_apply_fit=gaussian_augmentation_apply_fit,
                 gaussian_augmentation_apply_predict=gaussian_augmentation_apply_predict,
                 seed=seed,
+                dataset_run_id=dataset_run_id,
+                dataset_tar_name=dataset_tar_name,
             )
         )
 
     return state
-
-
-# Update data dir path if user is applying defense over image artifacts.
-def download_image_archive(
-    run_id: str, archive_path: str, destination_path: Optional[str] = None
-) -> str:
-    client: MlflowClient = MlflowClient()
-    image_archive_path: str = client.download_artifacts(
-        run_id=run_id, path=archive_path, dst_path=destination_path
-    )
-    LOGGER.info(
-        "Image archive downloaded",
-        run_id=run_id,
-        storage_path=archive_path,
-        dst_path=image_archive_path,
-    )
-    return image_archive_path
 
 
 def init_guassian_augmentation_flow() -> Flow:
@@ -264,6 +230,8 @@ def init_guassian_augmentation_flow() -> Flow:
             gaussian_augmentation_apply_fit,
             gaussian_augmentation_apply_predict,
             seed,
+            dataset_run_id,
+            dataset_tar_name,
         ) = (
             Parameter("testing_dir"),
             Parameter("image_size"),
@@ -277,6 +245,8 @@ def init_guassian_augmentation_flow() -> Flow:
             Parameter("gaussian_augmentation_apply_fit"),
             Parameter("gaussian_augmentation_apply_predict"),
             Parameter("seed"),
+            Parameter("dataset_run_id"),
+            Parameter("dataset_tar_name"),
         )
         seed, rng = pyplugs.call_task(
             f"{_PLUGINS_IMPORT_PATH}.random", "rng", "init_rng", seed=seed
@@ -310,7 +280,19 @@ def init_guassian_augmentation_flow() -> Flow:
                 dataset_seed=dataset_seed,
             ),
         )
-
+        adv_tar_path = pyplugs.call_task(
+            f"{_PLUGINS_IMPORT_PATH}.artifacts",
+            "mlflow",
+            "download_all_artifacts_in_run",
+            run_id=dataset_run_id,
+            artifact_path=dataset_tar_name,
+        )
+        extract_tarfile_results = pyplugs.call_task(
+            f"{_PLUGINS_IMPORT_PATH}.artifacts",
+            "utils",
+            "extract_tarfile",
+            filepath=adv_tar_path,
+        )
         distance_metrics_list = pyplugs.call_task(
             f"{_PLUGINS_IMPORT_PATH}.metrics",
             "distance",
@@ -332,7 +314,7 @@ def init_guassian_augmentation_flow() -> Flow:
             sigma=gaussian_augmentation_sigma,
             apply_fit=gaussian_augmentation_apply_fit,
             apply_predict=gaussian_augmentation_apply_predict,
-            upstream_tasks=[make_directories_results],
+            upstream_tasks=[make_directories_results, extract_tarfile_results],
         )
         log_evasion_dataset_result = pyplugs.call_task(  # noqa: F841
             f"{_PLUGINS_IMPORT_PATH}.artifacts",
