+++ examples/tensorflow-backdoor-poisoning/src/init_model.py
@@ -148,13 +148,6 @@ def init_model(
         seed=seed,
     )
 
-    mlflow.autolog()
-
-    if imagenet_preprocessing:
-        rescale = 1.0
-    else:
-        rescale = 1.0 / 255
-
     with mlflow.start_run() as active_run:
         flow: Flow = init_train_flow()
         state = flow.run(
@@ -168,7 +161,6 @@ def init_model(
                 learning_rate=learning_rate,
                 optimizer_name=optimizer,
                 imagenet_preprocessing=imagenet_preprocessing,
-                rescale=rescale,
                 seed=seed,
             )
         )
@@ -187,7 +179,6 @@ def init_train_flow() -> Flow:
             learning_rate,
             optimizer_name,
             imagenet_preprocessing,
-            rescale,
             seed,
         ) = (
             Parameter("active_run"),
@@ -199,7 +190,6 @@ def init_train_flow() -> Flow:
             Parameter("learning_rate"),
             Parameter("optimizer_name"),
             Parameter("imagenet_preprocessing"),
-            Parameter("rescale"),
             Parameter("seed"),
         )
         seed, rng = pyplugs.call_task(
@@ -217,6 +207,12 @@ def init_train_flow() -> Flow:
             "init_tensorflow",
             seed=tensorflow_global_seed,
         )
+        rescale = pyplugs.call_task(
+            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_poisoning_plugins",
+            "datasetup",
+            "select_rescale_value",
+            imagenet_preprocessing=imagenet_preprocessing,
+        )
 
         log_mlflow_params_result = pyplugs.call_task(  # noqa: F841
             f"{_PLUGINS_IMPORT_PATH}.tracking",
@@ -229,7 +225,7 @@ def init_train_flow() -> Flow:
             ),
         )
         optimizer = pyplugs.call_task(
-            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_poisoning_plugins",
+            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.evaluation",
             "tensorflow",
             "get_optimizer",
             optimizer=optimizer_name,
@@ -237,14 +233,14 @@ def init_train_flow() -> Flow:
             upstream_tasks=[init_tensorflow_results],
         )
         metrics = pyplugs.call_task(
-            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_poisoning_plugins",
+            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.evaluation",
             "tensorflow",
             "get_performance_metrics",
             metrics_list=PERFORMANCE_METRICS,
             upstream_tasks=[init_tensorflow_results],
         )
         callbacks_list = pyplugs.call_task(
-            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_poisoning_plugins",
+            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.evaluation",
             "tensorflow",
             "get_model_callbacks",
             callbacks_list=CALLBACKS,
@@ -283,7 +279,7 @@ def init_train_flow() -> Flow:
             training=False,
         )
         classifier_performance_metrics = pyplugs.call_task(
-            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_poisoning_plugins",
+            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.evaluation",
             "tensorflow",
             "evaluate_metrics_tensorflow",
             classifier=classifier,
